refs #21630 Refactored code for MuonMaxent
